+++ level1/p11_linkedList/linkedList.cpp
@@ -1,44 +1,8 @@
-#include <iostream>
 #include <sstream>
 
-template <typename T>
-class linearList {  // ADT 线性表
-   public:
-    virtual ~linearList(){};
-    virtual bool empty() const = 0;
-    virtual int size() const = 0;
-    virtual T& get(int theIndex) const = 0;
-    virtual int indexOf(const T& theElement, int theIndex) const = 0;
-    virtual void erase(int theIndex) = 0;
-    virtual void insert(int theIndex, const T& theElement) = 0;
-    virtual void output(std::ostream& out) const = 0;
-    virtual void clear() = 0;
-    virtual void push_back(const T& theElement) = 0;
-    virtual void pop_back() = 0;
-};
-
-class illegalParameter {
-   public:
-    illegalParameter() : message("Illegal parameter value") {}
-    illegalParameter(std::string theMessage) { message = theMessage; }
-    void outputMessage() { std::cout << message << std::endl; }
-
-   private:
-    std::string message;
-};
-
-template <typename T>
-class linkedNode {  //链表节点
-   public:
-    linkedNode() {}
-    linkedNode(T theElement) { element = theElement; }
-    linkedNode(T theElement, linkedNode* theNext) {
-        element = theElement;
-        next = theNext;
-    }
-    T element;
-    linkedNode* next;
-};
+#include "illegalParameter.h"
+#include "linearList.h"
+#include "linkedNode.h"
 
 template <typename T>
 class linkedList : public linearList<T> {  //链表
